Allow local datasets to be used without the DataController caching anything
Adds a 'caching' parameter to ModelController.run().
If True (default):
Data from the hydromodel is read and written
to a cache file by the DataController.  Particles read from that cache
file.  Useful for DAP datasets so you only download the same data once.
If False:
Data is read DIRECTLY from the hydromodel by the Particle, and the
DataController really does nothing.  Useful for local datasets.